Moved header to its own component, removed StaticQuery for test on Header, updated package.json from default Gatsby starter and README for running tests.
