+++ README.md
@@ -1,6 +1,6 @@
 <p align="center">
   <a href="http://www.freefrom.org/">
-    <img alt="FreeFrom" src="http://static1.squarespace.com/static/56a24df4d8af10a5072bed7c/t/56a2631b841aba12ab7f66d0/1562878613622/?format=1500w" width="60" />
+    <img alt="FreeFrom" src="http://static1.squarespace.com/static/56a24df4d8af10a5072bed7c/t/56a2631b841aba12ab7f66d0/1562878613622/?format=1500w" width="200" />
   </a>
 </p>
 
@@ -12,7 +12,7 @@ This is the front end web for FreeFrom. It's back end counterpart is located [he
 
 ## Tech
 - Gatsby/React Front End
-- Ruby Back End
+- Ruby/Rails Back End
 - Postgres DB
 
 ## Requirements to run Front End
@@ -22,58 +22,23 @@ This is the front end web for FreeFrom. It's back end counterpart is located [he
 ## 🚀 Quick start
 
 1. Install node_modules `npm install`
-2. Build/run gatsby `gatsby develop`
+2. Build/run gatsby `npm run develop`
 
   Your site is now running at `http://localhost:8000`!
 
   _Note: You'll also see a second link: _`http://localhost:8000/___graphql`_. This is a tool you can use to experiment with querying your data. Learn more about using this tool in the [Gatsby tutorial](https://www.gatsbyjs.org/tutorial/part-five/#introducing-graphiql).
 
-## 🧐 What's inside?
+## 🔬 Running Tests
+We use Jest for running unit tests. Refer to Gatsby guide ([Link](https://www.gatsbyjs.org/docs/unit-testing)) for some best practices on generating tests on Gatsby.
 
-A quick look at the top-level files and directories you'll see in a Gatsby project.
+1. To run tests: `npm run test`
+2. If tests fail due to the snapshots not being updated, run: `npm test -- -u`
 
-    .
-    ├── node_modules
-    ├── src
-    ├── .gitignore
-    ├── .prettierrc
-    ├── gatsby-browser.js
-    ├── gatsby-config.js
-    ├── gatsby-node.js
-    ├── gatsby-ssr.js
-    ├── LICENSE
-    ├── package-lock.json
-    ├── package.json
-    └── README.md
-
-1.  **`/node_modules`**: This directory contains all of the modules of code that your project depends on (npm packages) are automatically installed.
-
-2.  **`/src`**: This directory will contain all of the code related to what you will see on the front-end of your site (what you see in the browser) such as your site header or a page template. `src` is a convention for “source code”.
-
-3.  **`.gitignore`**: This file tells git which files it should not track / not maintain a version history for.
-
-4.  **`.prettierrc`**: This is a configuration file for [Prettier](https://prettier.io/). Prettier is a tool to help keep the formatting of your code consistent.
-
-5.  **`gatsby-browser.js`**: This file is where Gatsby expects to find any usage of the [Gatsby browser APIs](https://www.gatsbyjs.org/docs/browser-apis/) (if any). These allow customization/extension of default Gatsby settings affecting the browser.
-
-6.  **`gatsby-config.js`**: This is the main configuration file for a Gatsby site. This is where you can specify information about your site (metadata) like the site title and description, which Gatsby plugins you’d like to include, etc. (Check out the [config docs](https://www.gatsbyjs.org/docs/gatsby-config/) for more detail).
-
-7.  **`gatsby-node.js`**: This file is where Gatsby expects to find any usage of the [Gatsby Node APIs](https://www.gatsbyjs.org/docs/node-apis/) (if any). These allow customization/extension of default Gatsby settings affecting pieces of the site build process.
-
-8.  **`gatsby-ssr.js`**: This file is where Gatsby expects to find any usage of the [Gatsby server-side rendering APIs](https://www.gatsbyjs.org/docs/ssr-apis/) (if any). These allow customization of default Gatsby settings affecting server-side rendering.
-
-9.  **`LICENSE`**: This project is licensed under the MIT license.
-
-10. **`package-lock.json`** (See `package.json` below, first). This is an automatically generated file based on the exact versions of your npm dependencies that were installed for your project. **(You won’t change this file directly).**
-
-11. **`package.json`**: A manifest file for Node.js projects, which includes things like metadata (the project’s name, author, etc). This manifest is how npm knows which packages to install for your project.
-
-12. **`README.md`**: A text file containing useful reference information about this project.
-
+Optionally, you can run tests as files change by running: `npm test -- --watch`
 
 ## 🎓 Learning Gatsby
 
-Looking for more guidance? Full documentation for Gatsby lives [on the website](https://www.gatsbyjs.org/). Here are some places to start:
+Looking for more guidance on Gatsby? Full documentation for Gatsby lives [on the website](https://www.gatsbyjs.org/). Here are some places to start:
 
 - **For most developers, we recommend starting with our [in-depth tutorial for creating a site with Gatsby](https://www.gatsbyjs.org/tutorial/).** It starts with zero assumptions about your level of ability and walks through every step of the process.
