+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "gatsby-starter-default",
+  "name": "freefrom-compensation-web",
   "private": true,
-  "description": "A simple starter to get up and developing quickly with Gatsby",
+  "description": "FreeFrom Quiz",
   "version": "0.1.0",
-  "author": "Kyle Mathews <mathews.kyle@gmail.com>",
+  "author": "RagTag",
   "dependencies": {
     "gatsby": "^2.13.18",
     "gatsby-image": "^2.2.6",
@@ -40,9 +40,9 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+    "url": "https://github.com/RagtagOpen/freefrom-compensation-web"
   },
   "bugs": {
-    "url": "https://github.com/gatsbyjs/gatsby/issues"
+    "url": "https://github.com/RagtagOpen/freefrom-compensation-web/issues"
   }
 }
