+++ src/components/Header/header.test.js
@@ -1,0 +1,13 @@
+import React from "react"
+import renderer from "react-test-renderer"
+
+import Header from "./header"
+
+describe("Header", () => {
+  it("renders correctly", () => {
+    const tree = renderer
+      .create(<Header siteTitle="FreeFrom" />)
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+})
